// js/app/src/components/agent/__generated__/AgentSessionsResource_sessions.graphql.ts
export type AgentSessionsResource_sessions$data = {
  readonly agentSessions: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly createdAt: string;
        readonly id: string;
        readonly isActive: boolean;
        readonly isTemporary: boolean;
        readonly title: string;
        readonly updatedAt: string;
        readonly " $fragmentSpreads": FragmentRefs<"EditAgentSessionTitleDialog_session">;
      };
    }>;
  };
  readonly " $fragmentType": "AgentSessionsResource_sessions";
};
export type AgentSessionsResource_sessions$key = {
  readonly " $data"?: AgentSessionsResource_sessions$data;
  readonly " $fragmentSpreads": FragmentRefs<"AgentSessionsResource_sessions">;
};